add and change texts

+++ bot/bot/handlers/applicant/callbacks.py
@@ -4,6 +4,8 @@ from aiogram.fsm.context import FSMContext
 from bot import keyboards
 from bot.services import users_service
 from bot.states import RegisterState
+from bot.misc.configuration import conf
+
 from .messages import menu
 
 callbacks_router = Router()
@@ -32,9 +34,14 @@ async def get_city(callback: types.CallbackQuery, state: FSMContext):
     await state.clear()
 
     await callback.message.answer(
-        '[Приветственное видео](https://drive.google.com/file/d/1v_EEKgyw5nruL6MWcj7zIYTHnok6c0Cp/view?usp=share_link)',
-        parse_mode='MARKDOWN',
-        reply_markup=keyboards.applicant.get_continue_keyboard('tg_channel')
+        'Привет, меня зовут Тимур! \n'
+        '\nЯ генеральный директор федеральной сети салонов красоты "ЦирюльникЪ" 💅🏽\n'
+        '\nЗаписал для тебя видео, в котором хочу поближе познакомить тебя с компанией 🤝',
+        reply_markup=keyboards.applicant.get_continue_url_keyboard(
+            'Посмотреть видео',
+            'https://drive.google.com/file/d/1v_EEKgyw5nruL6MWcj7zIYTHnok6c0Cp/view?usp=share_link',
+            'tg_channel'
+        )
     )
 
 
@@ -42,23 +49,26 @@ async def get_city(callback: types.CallbackQuery, state: FSMContext):
 async def send_tg_channel(callback: types.CallbackQuery):
     await callback.message.delete()
     await callback.message.answer(
-        '[Давай сверимся что вы подписаны на наш канал с вакансиями]('
-        'https://drive.google.com/file/d/1alhaR1KPaZzLs_xexPmhUxK9DKgnQF54/view?usp=share_link)\n'
-        '"Ссылка на канал"',
-        parse_mode='MARKDOWN',
-        reply_markup=keyboards.applicant.get_continue_keyboard('salon')
+        'Давай сверимся, что ты подписан на наш канал с вакансиями 📱',
+        reply_markup=keyboards.applicant.get_channel_keyboard(conf.bot.channel_url, 'salon')
     )
 
 
 @callbacks_router.callback_query(F.data == 'salon')
 async def send_tg_channel(callback: types.CallbackQuery):
-    await callback.message.delete()
-    await callback.message.answer(
-        '[Давайте посмотрим на наши салоны изнутри]('
-        'https://drive.google.com/file/d/1sxSb2h8lptIsjVe8-psOHR1Cnks9tId5/view?usp=share_link)\n',
-        parse_mode='MARKDOWN',
-        reply_markup=keyboards.applicant.get_continue_keyboard('city')
-    )
+    user_channel_status = await callback.bot.get_chat_member(chat_id=conf.bot.channel, user_id=callback.from_user.id)
+    if user_channel_status != 'left':
+        await callback.message.delete()
+        await callback.message.answer(
+            'Давай посмотрим на наши салоны изнутри 📷',
+            reply_markup=keyboards.applicant.get_continue_url_keyboard(
+                'Посмотреть',
+                'https://drive.google.com/file/d/1sxSb2h8lptIsjVe8-psOHR1Cnks9tId5/view?usp=share_link',
+                'city',
+            )
+        )
+    else:
+        await send_tg_channel(callback)
 
 
 @callbacks_router.callback_query(F.data == 'city')
@@ -66,9 +76,12 @@ async def send_city(callback: types.CallbackQuery):
     await callback.message.delete()
     user = await users_service.get_by_tg_id(callback.from_user.id)
     await callback.message.answer(
-        f'[{user["city"]} видео]({cities_videos[user['city']]})',
-        parse_mode='MARKDOWN',
-        reply_markup=keyboards.applicant.get_continue_keyboard('menu')
+        'Видео про твой город 📍',
+        reply_markup=keyboards.applicant.get_continue_url_keyboard(
+            'Посмотреть',
+            cities_videos[user['city']],
+            'city',
+        )
     )
